requirements.txt
Rpi.bme280
azure-iot-device
RPi.GPIO